% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprints.R
\name{fingerprints}
\alias{fingerprints}
\title{Fingerprint input}
\usage{
fingerprints(fps, format = "full")
}
\arguments{
\item{fps}{Character vector of fingerprints (optionally named)}

\item{format}{Format of fingerprints. Can either be "full" or "rle" (see Details).}
}
\value{
Character vector of fingerprints suitable as \code{MorganFPS$new()} input
}
\description{
Make character vector of fingerprints.
}
\details{
Run-length encoded (RLE) representations of fingerprints can be generated using RDkit:\preformatted{AllChem.GetMorganFingerprintAsBitVect(
  inchi.MolFromInchi("InChI=1S/C17H18N6/c18-7-5-15(12-3-1-2-4-12)23-10-13(9-22-23)16-14-6-8-19-17(14)21-11-20-16/h6,8-12,15H,1-5H2,(H,19,20,21)/t15-/m1/s1"),2,nBits=1024
).ToBinary().hex()
}

"full" representations are hex-encoded versions of the binary output of RDkit:\preformatted{AllChem.GetMorganFingerprintAsBitVect(
  inchi.MolFromInchi("InChI=1S/C17H18N6/c18-7-5-15(12-3-1-2-4-12)23-10-13(9-22-23)16-14-6-8-19-17(14)21-11-20-16/h6,8-12,15H,1-5H2,(H,19,20,21)/t15-/m1/s1"),2,nBits=1024
).ToBitString()
}
}
\examples{
fingerprints(
 c(
  "e0ffffff000400001a00000026582eb62a09002c2a620c18309638069c0644327826440c324e3e98",
  "e0ffffff0004000037000000043c221a0c0e0c044a140ede321e3e2c06000e104c14362c040c206a3a76022a06401c1800540402021e002a0a00183e1640101a462e1208",
  "e0ffffff000400003200000002042c04182a1c125a0006ae8e2a00140618060622620e365c92307e0a023c021478041e02320e0c04282046000e10124c3208"
 ),
 format = "rle"
)
fingerprints(
 c(
  "02000000000000000000010000100000000000000000000000000000020000000000000802000000000000000000000420008000000000000008000000000000008000000000000200000200000000000004000800000000000000000000000000000000800000000000000000080000180000408000000000802000000000000000000002000080010200000000000401008800008000000001000000000000800000000000000000202000100000000001000000000000000100001000000000000000202020000000000000000000020000000000000000004000010000000000000000000180000002000000002000000000040000000000000200000000",
  "0200040000000000008081010000800000000000000000000000004008200000040000080000000000400000010000042000000000000000000000000000000010000000000000000000006100000000000400000000a000000024000100000000408080a00c000000000000000a0204000000488000208000000001000000080080000002040040010000044000000201090000008000000000600000200000000000000020000001200400140080000404000000000080080580000000000000000010000008008420010000001000000000000000001021004004000000c000040000000000c0000002000000802000000080001400004000000004000000",
 ),
 format = "full"
)
}
