% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeCollection}
\alias{makeCollection}
\title{Top-level function that generates and returns a MorganFPS object}
\arguments{
\item{vhx}{Vector of hexadecimal strings}
}
\value{
A MorganFPS object
}
\description{
Top-level function that generates and returns a MorganFPS object
}
