% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MorganFPS}
\alias{MorganFPS}
\title{Morgan fingerprints collection}
\description{
Efficient structure for storing a set of Morgan fingerprints
}
\section{Fields}{

\describe{
\item{\code{new}}{Construct new fingerprint dataset

Accepts either a vector of fingerprints in hexadecimal
format or a path to a binary file of fingerprints

The vector of fingerprints passed to the constructor can optionally be
named. Names need to be coercible to integers. The names can then be used
to refer to fingerprints in all functions using this object.
\itemize{
\item Parameter fingerprints - Character vector of fingerprints,
optionally wrapped in \code{\link[=fingerprints]{fingerprints()}}, or path to fingerprint file
saved using \code{save_file()}.
\item Parameter: from_file (default FALSE) - Set true to load from file
}}

\item{\code{tanimoto}}{similarity between fingerprints i and j \itemize{
\item Parameters: i, j - integer labels of two fingerprints
\item Returns: scalar numeric - Tanimoto similarity
}}

\item{\code{tanimoto_all}}{similarity between fingerprint i and all others \itemize{
\item Parameter: i - integer label of fingerprint
\item Returns: Dataframe with columns "id" and "similarity"
}}

\item{\code{tanimoto_threshold}}{similarity of all NxN combinations of fingerprints
above the given threshold \itemize{
\item Parameter: threshold - numeric threshold between 0 and 1
\item Returns: Dataframe with columns "id_1", "id_2", and "similarity"
}}

\item{\code{tanimoto_subset}}{similarity of a set of fingerprints against another set,
or all fingerprints in the collection when j is NULL \itemize{
\item Parameters: i, j - vectors of fingerprint labels. j can be NULL.
\item Returns: Dataframe with columns "id_1", "id_2", and "similarity"
}}

\item{\code{tanimoto_ext}}{similarity between given fingerprint and all
fingerprints in the collection \itemize{
\item Parameter: s - Fingerprint, optionally wrapped in \code{\link[=fingerprints]{fingerprints()}}
to specify encoding
\item Returns: Dataframe with columns "id" and "similarity"
}}

\item{\code{save_file}}{Save fingerprints to file in binary format \itemize{
\item Parameter: path - Path to location where fingerprints will be stored
\item Parameter: compression_level (default 3) - Optional integer between
0 and 22 specifying the level of compression used. Higher values produce
smaller files at the cost of slowing down writing
}}

\item{\code{n}}{number of fingerprints}

\item{\code{size}}{number of bytes used to store the fingerprints}
}}

