% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MorganMap}
\alias{MorganMap}
\title{Morgan fingerprint collection for identity checking}
\description{
Efficient structure for checking identity of Morgan fingerprints
}
\section{Fields}{

\describe{
\item{\code{new}}{Construct new fingerprint dataset

The vector of fingerprints passed to the constructor can optionally be
named. Names need to be coercible to integers. The names can then be used
to refer to fingerprints in all functions using this object.
\itemize{
\item Parameter fingerprints - Character vector of fingerprints,
optionally wrapped in \code{\link[=fingerprints]{fingerprints()}}
}}

\item{\code{find_matches}}{Find fingerprints in the collection that are identical
to the given fingerprints \itemize{
\item Parameter fingerprints - Character vector of fingerprints,
optionally wrapped in \code{\link[=fingerprints]{fingerprints()}}
}}
}}

