% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tanimoto}
\alias{tanimoto}
\title{Tanimoto similarity between two Morgan fingerprints}
\usage{
tanimoto(s1, s2)
}
\arguments{
\item{s1, s2}{Two fingerprints, optionally each wrapped in \code{\link[=fingerprints]{fingerprints()}}}
}
\value{
Jaccard similarity over the bits representing individual keys
}
\description{
Computes Tanimoto similarity between two hexadecimal strings
}
