% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MorganFPS}
\alias{MorganFPS}
\title{Morgan fingerprints collection}
\description{
Efficient structure for storing a set of Morgan fingerprints
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor}

\item{\code{tanimoto}}{(i,j) similarity between fingerprints i and j}

\item{\code{tanimoto_all}}{(i) similarity between fingerprint i and all others}

\item{\code{tanimoto_ext}}{(s) similarity between external hexadecimal string s and all
fingerprints in the collection}

\item{\code{size}}{number of bytes used to store the fingerprints}
}}

